Add docker role, update inventory.py, refactoring
Add a role to install dokker and docker-compose
The inventory.py script gets ansible_ssh_host via libvirt
