+++ inventory/group_vars/all.yaml
@@ -1,0 +1,2 @@
+default_user: ahmed
+default_group: ahmed
