+++ inventory/hosts.yaml
@@ -1,8 +1,8 @@
 ---
-prometheus_centos_1:
+prometheus-centos-1:
   groups: [ monitoring, prometheus ]
-  hostvars: { ansible_ssh_host: 192.168.122.5, ansible_ssh_user: ahmed }
+  hostvars: { ansible_ssh_user: ahmed, os_type: rhel }
 
-prometheus_ubuntu_1:
-  groups: [ monitoring, prometheus ]
-  hostvars: { ansible_ssh_host: 192.168.122.95, ansible_ssh_user: ahmed }
+proxmox-ub-01:
+  groups: [ proxmox ]
+  hostvars: { ansible_ssh_user: ahmed, os_type: debian }
