--- inventory/inventory.ini
@@ -1,3 +1,0 @@
-[prometheus]
-prometheus_centos ansible_host=192.168.122.5 ansible_user=ahmed
-prometheus_ubuntu ansible_host=192.168.122.95 ansible_user=ahmed
