+++ roles/docker/tasks/ubuntu.yaml
@@ -1,0 +1,42 @@
+- name: Uninstall all conflicting packages
+  apt:
+    name: "{{ item }}"
+    state: absent
+    update_cache: yes
+  loop:
+    - docker.io
+    - docker-doc
+    - docker-compose
+    - docker-compose-v2
+    - podman-docker
+    - containerd
+    - runc
+
+- name: Add docker key
+  get_url:
+    url: https://download.docker.com/linux/ubuntu/gpg
+    dest: /etc/apt/keyrings/docker.asc
+    mode: 0644
+
+- name: Get distro arch
+  shell: dpkg --print-architecture
+  register: distro_arch
+
+- name: Add docker repository
+  apt_repository:
+    repo: "deb [arch={{ distro_arch.stdout }} signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu   {{ ansible_distribution_release }} stable"
+    state: present
+    update_cache: yes
+    filename: docker
+    mode: 0644
+
+- name: Install docker
+  apt:
+    name: "{{ item }}"
+    state: present
+  loop:
+    - docker-ce
+    - docker-ce-cli
+    - containerd.io
+    - docker-buildx-plugin
+    - docker-compose-plugin
