+++ roles/ubuntu-common/tasks/main.yaml
@@ -3,5 +3,3 @@
     update_cache: yes
     upgrade: yes
     state: latest
-  become: yes
-  become_user: root
